Rename authors-list constants file
